AS 4.1 compatibility (fixes #36)
