+++ globallydynamic-studio-plugin/build.gradle
@@ -1,6 +1,6 @@
 plugins {
-    id 'org.jetbrains.intellij' version '0.4.21'
-    id 'org.jetbrains.kotlin.jvm' version '1.3.72'
+    id 'org.jetbrains.intellij' version '0.5.0'
+    id 'org.jetbrains.kotlin.jvm' version '1.4.0'
 }
 
 apply from: 'deps.gradle'
@@ -36,9 +36,7 @@ compileTestKotlin {
 
 intellij {
     plugins 'android'
-    type 'IC'
-    version '193.6911.18'
-//    localPath "${System.getenv("ANDROID_STUDIO_PATH")}"
+    localPath "${System.getenv("ANDROID_STUDIO_PATH")}"
 }
 
 publishPlugin {
@@ -47,7 +45,7 @@ publishPlugin {
 }
 
 patchPluginXml {
-    // 1.0
+    // 1.0.0
     changeNotes """
       Compatible with Android Studio 4.0
       
@@ -60,4 +58,9 @@ patchPluginXml {
       
       <p>Find the documentation <a href="https://globallydynamic.io">here</a></p>
       """
+
+    // 1.1.0
+    """
+    Android Studio 4.1 compatibility
+    """
 }
