+++ globallydynamic-studio-plugin/deps.gradle
@@ -1,5 +1,5 @@
 def versions = [
-        kotlin                : '1.3.72',
+        kotlin                : '1.4.0',
         globallydynamic_server: '1.0.0',
         junit                 : '5.5.2',
         junit_platform        : '1.5.2',
