+++ globallydynamic-studio-plugin/gradle.properties
@@ -1,5 +1,5 @@
 GROUP=com.jeppeman.globallydynamic.idea
-VERSION_NAME=1.0.0
+VERSION_NAME=1.1.0
 
 POM_DESCRIPTION=GloballyDynamic Studio - Embeds a GloballyDynamic server into Android Studio that apps will download their splits from.
