+++ globallydynamic-studio-plugin/src/main/kotlin/com/jeppeman/globallydynamic/idea/PostProjectSyncStep.kt
@@ -2,9 +2,11 @@ package com.jeppeman.globallydynamic.idea
 
 import com.android.annotations.VisibleForTesting
 import com.android.tools.idea.gradle.project.sync.setup.post.ProjectSetupStep
+import com.android.tools.idea.run.AndroidRunConfigurationBase
+import com.intellij.execution.RunManagerEx
+import com.intellij.execution.configurations.RunConfiguration
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleManager
-import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
 import com.jeppeman.globallydynamic.idea.extensions.hasGloballyDynamicEnabled
 
@@ -15,11 +17,49 @@ class PostProjectSyncStep : ProjectSetupStep() {
 
     @VisibleForTesting
     fun shouldInitializeServer(project: Project): Boolean =
-        ModuleManager.getInstance(project).modules.any(Module::hasGloballyDynamicEnabled)
+        getModuleManager(project).modules.any(Module::hasGloballyDynamicEnabled)
 
-    override fun setUpProject(project: Project, progressIndicator: ProgressIndicator?) {
+    @VisibleForTesting
+    fun getModuleManager(project: Project): ModuleManager = ModuleManager.getInstance(project)
+
+    @VisibleForTesting
+    fun getRunManagerX(project: Project): RunManagerEx = RunManagerEx.getInstanceEx(project)
+
+    @VisibleForTesting
+    fun getAndroidRunConfigurations(runManagerEx: RunManagerEx): List<RunConfiguration> =
+        runManagerEx.allConfigurationsList.filterIsInstance<AndroidRunConfigurationBase>()
+
+    @VisibleForTesting
+    fun getGloballyDynamicBuildPreparationProvider() = GloballyDynamicBuildPreparationProvider()
+
+    @VisibleForTesting
+    fun shouldAddBuildPreparationTask(module: Module): Boolean = module.hasGloballyDynamicEnabled
+
+    override fun setUpProject(project: Project) {
         if (shouldInitializeServer(project)) {
             getServerManager(project).start()
+        }
+
+        getModuleManager(project).modules.forEach { module ->
+            if (shouldAddBuildPreparationTask(module)) {
+                val runManagerEx = getRunManagerX(module.project)
+                val androidRunConfigurations = getAndroidRunConfigurations(runManagerEx)
+
+                val globallyDynamicBuildPreparationProvider = getGloballyDynamicBuildPreparationProvider()
+                androidRunConfigurations.forEach { androidRunConfiguration ->
+                    val tasks = runManagerEx.getBeforeRunTasks(androidRunConfiguration)
+                    val globallyDynamicBuildPreparationTasks = tasks.filterIsInstance<GloballyDynamicBuildPreparationTask>()
+                    if (globallyDynamicBuildPreparationTasks.isEmpty()) {
+                        val task = globallyDynamicBuildPreparationProvider.createTask(androidRunConfiguration)
+                        tasks.add(0, task)
+                        runManagerEx.setBeforeRunTasks(androidRunConfiguration, tasks)
+                    } else {
+                        val otherTasks = tasks.filter { !globallyDynamicBuildPreparationTasks.contains(it) }
+                        val sortedTasks = listOf(globallyDynamicBuildPreparationTasks.first()) + otherTasks
+                        runManagerEx.setBeforeRunTasks(androidRunConfiguration, sortedTasks)
+                    }
+                }
+            }
         }
     }
 }
