+++ globallydynamic-studio-plugin/src/main/resources/META-INF/plugin.xml
@@ -11,6 +11,7 @@
         A server log tool window is also included for debugging purposes.
     </description>
 
+    <depends>com.intellij.modules.androidstudio</depends>
     <depends>org.jetbrains.android</depends>
 
     <extensions defaultExtensionNs="com.intellij">
@@ -24,7 +25,6 @@
     </extensions>
 
     <extensions defaultExtensionNs="com.android.gradle.sync">
-        <postSyncModuleSetupStep implementation="com.jeppeman.globallydynamic.idea.PostModuleSyncStep"/>
         <postSyncProjectSetupStep implementation="com.jeppeman.globallydynamic.idea.PostProjectSyncStep"/>
     </extensions>
